created ReservationsBase module
