+++ app/models/reservation_validations.rb
@@ -1,12 +1,5 @@
 module ReservationValidations
 
-  def self.included(base)
-    base.belongs_to :equipment_model
-    base.belongs_to :reserver, :class_name => 'User'
-    base.attr_accessible :reserver, :reserver_id, :start_date, :due_date,
-                         :equipment_model_id
-  end
-
   ## Validations ##
 
   ## For individual reservations only
